amount transferred implemented across all transaction tables
