+++ explorer/client/src/components/transaction-card/TxForID.tsx
@@ -1,12 +1,8 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import {
-    type GetTxnDigestsResponse,
-    type ExecutionStatusType,
-    type TransactionKindName,
-} from '@mysten/sui.js';
-import { useState, useEffect, useContext, useMemo } from 'react';
+import { type GetTxnDigestsResponse } from '@mysten/sui.js';
+import { useState, useEffect, useContext } from 'react';
 
 import { NetworkContext } from '../../context';
 import {
@@ -14,31 +10,18 @@ import {
     getDataOnTxDigests,
 } from '../../utils/api/DefaultRpcClient';
 import { IS_STATIC_ENV } from '../../utils/envUtil';
-import { numberSuffix } from '../../utils/numberUtil';
 import { deduplicate } from '../../utils/searchUtil';
 import { findTxfromID, findTxDatafromID } from '../../utils/static/searchUtil';
-import { truncate } from '../../utils/stringUtils';
-import { timeAgo } from '../../utils/timeUtils';
 import ErrorResult from '../error-result/ErrorResult';
 import PaginationLogic from '../pagination/PaginationLogic';
 import TableCard from '../table/TableCard';
+import { type TxnData, genTableDataFromTxData } from './TxCardUtils';
 
 const TRUNCATE_LENGTH = 14;
 const ITEMS_PER_PAGE = 20;
 
 const DATATYPE_DEFAULT = {
     loadState: 'pending',
-};
-
-type TxnData = {
-    seq: number;
-    txId: string;
-    status: ExecutionStatusType;
-    kind: TransactionKindName | undefined;
-    From: string;
-    To?: string;
-    timestamp_ms?: number | null;
-    txGas: number;
 };
 
 type categoryType = 'address' | 'object';
@@ -55,79 +38,9 @@ const getTx = async (
 const viewFn = (results: any) => <TxForIDView showData={results} />;
 
 function TxForIDView({ showData }: { showData: TxnData[] | undefined }) {
-    // TODO: Ideally move this to a prop:
-    const hasTimeColumn = showData?.[0]?.timestamp_ms;
-
-    const tableData = useMemo(
-        () => ({
-            data: (showData ?? []).map((data) => ({
-                date: `${timeAgo(data.timestamp_ms, undefined, true)} ago`,
-                txTypes: {
-                    txTypeName: data.kind,
-                    status: data.status,
-                },
-                transactionId: [
-                    {
-                        url: data.txId,
-                        name: truncate(data.txId, 26),
-                        category: 'transactions',
-                        isLink: true,
-                        copy: false,
-                    },
-                ],
-                addresses: [
-                    {
-                        url: data.From,
-                        name: truncate(data.From, TRUNCATE_LENGTH),
-                        category: 'addresses',
-                        isLink: true,
-                        copy: false,
-                    },
-                    ...(data.To
-                        ? [
-                              {
-                                  url: data.To,
-                                  name: truncate(data.To, TRUNCATE_LENGTH),
-                                  category: 'addresses',
-                                  isLink: true,
-                                  copy: false,
-                              },
-                          ]
-                        : []),
-                ],
-                gas: numberSuffix(data.txGas),
-            })),
-            columns: [
-                ...(hasTimeColumn
-                    ? [
-                          {
-                              headerLabel: 'Time',
-                              accessorKey: 'date',
-                          },
-                      ]
-                    : []),
-                {
-                    headerLabel: 'TxType',
-                    accessorKey: 'txTypes',
-                },
-                {
-                    headerLabel: 'Transaction ID',
-                    accessorKey: 'transactionId',
-                },
-                {
-                    headerLabel: 'Addresses',
-                    accessorKey: 'addresses',
-                },
-                {
-                    headerLabel: 'Gas',
-                    accessorKey: 'gas',
-                },
-            ],
-        }),
-        [hasTimeColumn, showData]
-    );
-
     if (!showData || showData.length === 0) return null;
+
+    const tableData = genTableDataFromTxData(showData, TRUNCATE_LENGTH);
 
     return (
         <div data-testid="tx">
@@ -173,18 +86,8 @@ function TxForIDAPI({ id, category }: { id: string; category: categoryType }) {
             } else {
                 getDataOnTxDigests(network, transactions)
                     .then((data) => {
-                        const subData = data.map((el) => ({
-                            seq: el!.seq,
-                            txId: el!.txId,
-                            status: el!.status,
-                            kind: el!.kind,
-                            From: el!.From,
-                            To: el!.To,
-                            txGas: el!.txGas,
-                            timestamp_ms: el!.timestamp_ms,
-                        }));
                         setData({
-                            data: subData,
+                            data: data as TxnData[],
                             loadState: 'loaded',
                         });
                     })
